# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PromptHub API Documentation",
        "description": "API completa para gerenciamento de prompts de IA e agentes AI. Sistema que permite autenticacao via email/senha ou Google OAuth, cadastro e gerenciamento de agentes de IA personalizados, envio de prompts para OpenAI e Google Gemini.",
        "contact": {
            "name": "PromptHub Support",
            "email": "suporte@prompthub.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://dflix-prompthub-backend.a8zera.easypanel.host/api",
            "description": "Production Server (Easypanel)"
        },
        {
            "url": "https://api-prompthub.marcioleite.cloud/api",
            "description": "Production Server (Alternative)"
        }
    ],
    "paths": {
...
        "/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Registrar novo usuario",
                "description": "Cria uma nova conta de usuario com email e senha. Retorna o token JWT apos cadastro bem-sucedido.",
                "operationId": "register",
                "requestBody": {
                    "description": "Dados do usuario para registro",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nome completo do usuario",
                                        "type": "string",
                                        "example": "Joao Silva"
                                    },
                                    "email": {
                                        "description": "Email unico do usuario",
                                        "type": "string",
                                        "format": "email",
                                        "example": "joao@exemplo.com"
                                    },
                                    "password": {
                                        "description": "Senha do usuario (minimo 6 caracteres)",
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 6,
                                        "example": "senha123"
                                    },
                                    "password_confirmation": {
                                        "description": "Confirmacao da senha",
                                        "type": "string",
                                        "format": "password",
                                        "example": "senha123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User successfully registered"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Joao Silva"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "joao@exemplo.com"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-11-25T10:00:00.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validacao",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The email has already been taken."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login de usuario",
                "description": "Autentica usuario com email e senha. Retorna token JWT para autorizacao.",
                "operationId": "login",
                "requestBody": {
                    "description": "Credenciais do usuario",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "joao@exemplo.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senha123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Joao Silva"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "joao@exemplo.com"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais invalidas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid credentials"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout de usuario",
                "description": "Invalida o token JWT atual. Requer autenticacao.",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Logout realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully logged out"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Nao autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Obter usuario autenticado",
                "description": "Retorna as informacoes do usuario atualmente autenticado.",
                "operationId": "getAuthenticatedUser",
                "responses": {
                    "200": {
                        "description": "Dados do usuario autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Joao Silva"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "joao@exemplo.com"
                                        },
                                        "google_id": {
                                            "type": "string",
                                            "example": "1234567890",
                                            "nullable": true
                                        },
                                        "avatar": {
                                            "type": "string",
                                            "example": "https://lh3.googleusercontent.com/...",
                                            "nullable": true
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Nao autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/google": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Iniciar autenticacao Google OAuth",
                "description": "Retorna a URL de redirecionamento para autenticacao via Google OAuth",
                "operationId": "redirectToGoogle",
                "responses": {
                    "200": {
                        "description": "URL de redirecionamento do Google",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "url": {
                                            "type": "string",
                                            "example": "https://accounts.google.com/o/oauth2/v2/auth?client_id=..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao gerar URL do Google",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao redirecionar para Google"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/google/callback": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Callback do Google OAuth",
                "description": "Endpoint de callback do Google OAuth. Processa o codigo de autorizacao e redireciona para o frontend com o token JWT.",
                "operationId": "handleGoogleCallback",
                "parameters": [
                    {
                        "name": "code",
                        "in": "query",
                        "description": "Codigo de autorizacao retornado pelo Google",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redireciona para o frontend com token ou erro",
                        "headers": {
                            "Location": {
                                "description": "URL de redirecionamento (frontend/auth/callback?token={token} ou frontend/login?error=...)",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme. Example: 'Authorization: Bearer {token}'",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}